Revert "Fix"
This reverts commit e6c8daf23118ba48a5756cd301cde8b0d70f80df.

+++ source/HTcodec/encoder/termMELandVLCPackers.m
@@ -6,7 +6,7 @@ x_FF = uint8(255);
 state_MELPacker.MEL_tmp = bitshift(state_MELPacker.MEL_tmp, state_MELPacker.MEL_rem);
 
 MEL_mask = bitand(bitshift(x_FF, state_MELPacker.MEL_rem), x_FF); % if MEL_rem is 8, MEL_mask = 0
-VLC_mask = bitshift(x_FF, -(8 - int8(state_VLC.VLC_bits))); % if VLC_bits is 0, VLC_mask = 0
+VLC_mask = bitshift(x_FF, -(8 - state_VLC.VLC_bits)); % if VLC_bits is 0, VLC_mask = 0
 
 if bitor(MEL_mask, VLC_mask) == 0
     return; % last MEL byte cannot be FF, since then MEL rem would be < 8
